requirements.txt
streamlit==1.28.0
youtube-transcript-api==0.7.0
langchain==0.1.0
langchain-community==0.0.20
langchain-huggingface==0.0.1
langchain-text-splitters==0.0.1
faiss-cpu==1.7.4
python-dotenv==1.0.0